reorder for log_enc
